refactor: extract decode attn receipt validation
